moved individual steps of substitute appellant flow into own folders
